+++ constants.py
@@ -1,2 +1,6 @@
 from xypath import DOWN, UP, LEFT, RIGHT
-from bake import is_number, showtime
+from bake import showtime
+import overrides
+
+def is_number(cell):
+    return type(cell.value) in [int, float, long]
